Use ArgumentParser for cli command definition
ArgumentParser is a new library from Apple to parse CLI arguments. It enables llbuild-analyze to define its commands, options and flags in a declarative fashion.
More information here: https://swift.org/blog/argument-parser/

rdar://59846733
