+++ .gitignore
@@ -18,7 +18,7 @@
 #==============================================================================#
 # Commonly used name for build products, at the top level.
 /build
-/.build
+.build
 /Packages
 
 # Ignore Sphinx build products.
